% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c5h8_emissions.R
\name{C5H8_emissions}
\alias{C5H8_emissions}
\title{isoprene emissions}
\usage{
C5H8_emissions(Temperature, LDF, EF, LAI)
}
\arguments{
\item{Temperature}{(K)}

\item{LDF}{Light Dependent Fraction}

\item{EF}{Emission Factor}

\item{LAI}{Leaf Area Index}
}
\value{
isoprene emissions in molecules per cm2 per s
}
\description{
script to bring together all the functions to output isoprene emissions in molecules per cm2 per s
}

