% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_c5h8_emissions_molecules_cm2_s.R
\name{get_c5h8_emissions_molecules_cm2_s}
\alias{get_c5h8_emissions_molecules_cm2_s}
\title{isoprene emissions in molecules cm-2 s-1}
\usage{
get_c5h8_emissions_molecules_cm2_s(c5h8_emissions_ugm2h)
}
\arguments{
\item{c5h8_emissions_ugm2h}{isoprene emissions in ug m-2 h-1}
}
\value{
isoprene emissions in molecules per cm2 per s
}
\description{
convert C5H8 emissions in ug per m2 per h to molecules cm-2 s-1
return value of isoprene emissions in molecules per cm2 per s
}

