% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_c5h8_emissions_ug_m2_h.R
\name{get_c5h8_emissions_ug_m2_h}
\alias{get_c5h8_emissions_ug_m2_h}
\title{isoprene emissions in ug m-2 h-1}
\usage{
get_c5h8_emissions_ug_m2_h(gamma_c5h8, EF)
}
\arguments{
\item{gamma_c5h8}{gamma for isoprene emissions}

\item{EF}{emission factor}
}
\value{
isoprene emissions in ug per m2 per h
}
\description{
C5H8 emissions = gamma_c5h8 * sum(EFs) * Fractional grid coverage
return value of isoprene emissions in ug per m2 per h
}

