% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_A.R
\name{get_gamma_A}
\alias{get_gamma_A}
\title{gamma_A value
gamma_A = Fnew*Anew + Fgro*Agro + Fmat*Amat + Fsen*Asen
Output value of gamma_A}
\usage{
get_gamma_A()
}
\value{
gamma_A value
}
\description{
gamma_A value
gamma_A = Fnew*Anew + Fgro*Agro + Fmat*Amat + Fsen*Asen
Output value of gamma_A
}

