% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_c5h8.R
\name{get_gamma_c5h8}
\alias{get_gamma_c5h8}
\title{gamma value for isoprene emissions}
\usage{
get_gamma_c5h8(LAI, gamma_P, gamma_T, gamma_A, gamma_SM, gamma_CO2)
}
\arguments{
\item{LAI}{Leaf Area Index}

\item{gamma_P}{gamma due to radiation}

\item{gamma_T}{gamma due to temperature}

\item{gamma_A}{gamma due to age}

\item{gamma_SM}{gamma due to soil moisture}

\item{gamma_co2}{gamma due to CO2}
}
\value{
gamma_c5h8 value
}
\description{
gamma_c5h8 = C_ce * LAI * gamma_p * gamma_t * gamma_a * gamma_sm * gamma_co2
return value of final gamma for isoprene emissions
}

