% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_P.R
\name{get_gamma_P}
\alias{get_gamma_P}
\title{gamma_P value}
\usage{
get_gamma_P(LDF)
}
\arguments{
\item{LDF}{light dependent factor of BVOC}
}
\value{
gamma_P value
}
\description{
gamma_P = (1 - LDF) + LDF*gamma_LDF
gamma_LDF = C_p(alpha * PPFD / (1 + alpha^2*PPFD^2)^(1/2))
alpha = 0.004 - 0.0005ln(P_240)
C_p = 0.0468*exp(0.0005(P_24 - Ps))*P_240^0.6
return value of gamma due to PAR (Photosythetically Active Radiation)
}

