% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_sm.R
\name{get_gamma_sm}
\alias{get_gamma_sm}
\title{gamma_soil_moisture value
Output value of gamma_sm}
\usage{
get_gamma_sm()
}
\value{
gamma_sm value
}
\description{
gamma_soil_moisture value
Output value of gamma_sm
}

