% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_T.R
\name{get_gamma_T}
\alias{get_gamma_T}
\title{gamma_T value. gamma_t,LIF = 0 for isoprene}
\usage{
get_gamma_T(Temperature, LDF)
}
\arguments{
\item{Temperature}{(K)}

\item{LDF}{light dependent factor of BVOC}
}
\value{
gamma_T value
}
\description{
gamma_T = (1 - LDF)gamma_t,LIF + LDF*gamma_t,LDF
gamma_t,LDF = Eopt*C_T2*exp(C_T1*x) / (C_T2 - C_T1(1 - exp(C_T2*x)))
Eopt = C_eo*exp(0.05(T_24 - T_s))*exp(0.05(T_240 - T_s))
x = (1/T_opt - 1/T)/0.00831
T_opt = 313 + (0.6(T_240 - T_s))
return value of gamma due to temperature
}

