\name{MEGAN-package}
\alias{MEGAN-package}
\alias{MEGAN}
\docType{package}
\title{
\packageTitle{MEGAN}
}
\description{
\packageDescription{MEGAN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MEGAN}
\packageIndices{MEGAN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MEGAN}

Maintainer: \packageMaintainer{MEGAN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
